+++ Tools/vecml2sc.m
@@ -18,8 +18,8 @@ function [sc] = vecml2sc(Clm,Slm,lmax)
 
 l1 = 0;
 sc = zeros(lmax+1,2*lmax+1);
-for mm = 0:lmax
-    for ll = mm:lmax
+for mm = double(0:lmax)
+    for ll = double(mm:lmax)
         l1 = l1+1;
         if l1<= length(Clm)
             sc(ll+1,lmax+1+mm) = Clm(l1);
